+++ Gemfile
@@ -1,6 +1,9 @@
 source 'https://rubygems.org'
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
+gem 'pundit-resources',
+    github: 'better-together-org/pundit-resources'
+
 gemspec
 
 group :development, :test do
