+++ app/controllers/better_together/api_controller.rb
@@ -2,7 +2,12 @@ require_dependency 'jsonapi/resource_controller'
 
 module BetterTogether
   class ApiController < ::JSONAPI::ResourceController
-    include Pundit
+    include Pundit::ResourceController
+    # include Pundit
     protect_from_forgery with: :null_session
+
+    def current_user
+      nil
+    end
   end
 end
