+++ app/policies/better_together/community_membership_policy.rb
@@ -1,0 +1,14 @@
+
+module BetterTogether
+  class CommunityMembershipPolicy < ApplicationPolicy
+    def create?
+      true
+    end
+
+    class Scope < Scope
+      def resolve
+        scope.all
+      end
+    end
+  end
+end
