+++ app/policies/better_together/community_policy.rb
@@ -1,0 +1,14 @@
+
+module BetterTogether
+  class CommunityPolicy < ApplicationPolicy
+    def create?
+      true
+    end
+
+    class Scope < Scope
+      def resolve
+        scope.all
+      end
+    end
+  end
+end
