+++ app/policies/better_together/person_policy.rb
@@ -1,0 +1,14 @@
+
+module BetterTogether
+  class PersonPolicy < ApplicationPolicy
+    def create?
+      true
+    end
+
+    class Scope < Scope
+      def resolve
+        scope.all
+      end
+    end
+  end
+end
