+++ app/resources/better_together/api_resource.rb
@@ -4,6 +4,7 @@ module BetterTogether
   # Base JSONAPI serializer that sets common attrbutes
   class ApiResource < ::JSONAPI::Resource
     abstract
+    include Pundit::Resource
 
     attributes :bt_id, :created_at, :updated_at
   end
